refactor: 웹 모델 패키지를 website 도메인으로 정리
- 웹사이트 전용 엔티티가 사용하는 서비스와 DTO의 도메인 경계에 놓이도록 패키지를 이동

- QueryDSL static import와 테스트 fixture import를 함께 정리해 생성 클래스 경로 불일치를 방지

- 테이블명은 유지해 DB 스키마 변경 없이 Java 패키지 경계만 정리
