+++ src/test/java/gg/agit/konect/integration/domain/website/WebsiteApiTest.java
@@ -17,8 +17,8 @@ import org.junit.jupiter.api.Test;
 import gg.agit.konect.domain.club.enums.ClubCategory;
 import gg.agit.konect.domain.university.enums.Campus;
 import gg.agit.konect.domain.university.enums.UniversityRegion;
-import gg.agit.konect.domain.web.model.WebClub;
-import gg.agit.konect.domain.web.model.WebUniversity;
+import gg.agit.konect.domain.website.model.WebClub;
+import gg.agit.konect.domain.website.model.WebUniversity;
 import gg.agit.konect.support.IntegrationTestSupport;
 import gg.agit.konect.support.fixture.WebClubFixture;
 import gg.agit.konect.support.fixture.WebUniversityFixture;
